# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: HAIROU BTP Management Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/refresh:
    post:
      operationId: refreshToken
      tags: [auth]
      summary: Refresh JWT token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RefreshTokenRequest"
      responses:
        "200":
          description: Token refreshed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
...
  /expenses:
    get:
      operationId: listExpenses
      tags: [expenses]
      summary: List all expenses
      security:
        - bearerAuth: []
      parameters:
        - name: projectId
          in: query
          schema:
            type: integer
        - name: category
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of expenses
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Expense"
    post:
      operationId: createExpense
      tags: [expenses]
      summary: Create an expense
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateExpenseRequest"
      responses:
        "201":
          description: Expense created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Expense"
  /expenses/{id}:
    put:
      operationId: updateExpense
      tags: [expenses]
      summary: Update expense
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateExpenseRequest"
      responses:
        "200":
          description: Expense updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Expense"
    delete:
      operationId: deleteExpense
      tags: [expenses]
      summary: Delete expense
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Expense deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /expenses/{id}/validate:
    post:
      operationId: validateExpense
      tags: [expenses]
      summary: Validate or reject expense
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ValidateExpenseRequest"
      responses:
        "200":
          description: Expense validated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Expense"
  /messages:
    get:
      operationId: listMessages
      tags: [messages]
      summary: List messages
      security:
        - bearerAuth: []
      parameters:
        - name: projectId
          in: query
          schema:
            type: integer
        - name: recipientId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of messages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Message"
    post:
      operationId: sendMessage
      tags: [messages]
      summary: Send a message
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendMessageRequest"
      responses:
        "201":
          description: Message sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
  /messages/{id}/read:
    post:
      operationId: markMessageRead
      tags: [messages]
      summary: Mark message as read
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Message marked as read
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required:
        - error
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required:
        - success
    LoginRequest:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
      required:
        - email
        - password
    RefreshTokenRequest:
      type: object
      properties:
        refreshToken:
          type: string
      required:
        - refreshToken
    AuthResponse:
      type: object
      properties:
        token:
          type: string
        refreshToken:
          type: string
        user:
          $ref: "#/components/schemas/User"
      required:
        - token
        - refreshToken
        - user
    UserPermissions:
      type: object
      properties:
        canAddWorkers:
          type: boolean
        canDeleteWorkers:
          type: boolean
        canEditWorkers:
          type: boolean
        canAddExpenses:
          type: boolean
        canDeleteExpenses:
          type: boolean
        canAddProjects:
          type: boolean
        canViewFinances:
          type: boolean
        canManagePointage:
          type: boolean
      required:
        - canAddWorkers
        - canDeleteWorkers
        - canEditWorkers
        - canAddExpenses
        - canDeleteExpenses
        - canAddProjects
        - canViewFinances
        - canManagePointage
    User:
      type: object
      properties:
        id:
          type: integer
        email:
          type: string
        name:
          type: string
        role:
          type: string
          enum: [ADMIN, CHEF_CHANTIER, OUVRIER]
        isActive:
          type: boolean
        permissions:
          $ref: "#/components/schemas/UserPermissions"
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - email
        - name
        - role
        - isActive
...
    Expense:
      type: object
      properties:
        id:
          type: integer
        projectId:
          type: integer
        projectName:
          type: string
        title:
          type: string
        category:
          type: string
          enum: [MATERIAUX, EQUIPEMENTS, TRANSPORT, CARBURANT, NOURRITURE, AUTRE]
        amount:
          type: number
        date:
          type: string
          format: date
        supplier:
          type: string
        receiptUrl:
          type: string
        addedById:
          type: integer
        addedByName:
          type: string
        status:
          type: string
          enum: [EN_ATTENTE, APPROUVEE, REJETEE]
        adminComment:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - projectId
        - title
        - category
        - amount
        - date
        - status
    CreateExpenseRequest:
      type: object
      properties:
        projectId:
          type: integer
        title:
          type: string
        category:
          type: string
          enum: [MATERIAUX, EQUIPEMENTS, TRANSPORT, CARBURANT, NOURRITURE, AUTRE]
        amount:
          type: number
        date:
          type: string
          format: date
        supplier:
          type: string
        receiptUrl:
          type: string
      required:
        - projectId
        - title
        - category
        - amount
        - date
    UpdateExpenseRequest:
      type: object
      properties:
        title:
          type: string
        category:
          type: string
          enum: [MATERIAUX, EQUIPEMENTS, TRANSPORT, CARBURANT, NOURRITURE, AUTRE]
        amount:
          type: number
        date:
          type: string
          format: date
        supplier:
          type: string
    ValidateExpenseRequest:
      type: object
      properties:
        approved:
          type: boolean
        comment:
          type: string
      required:
        - approved
    Message:
      type: object
      properties:
        id:
          type: integer
        senderId:
          type: integer
        senderName:
          type: string
        recipientId:
          type: integer
        recipientName:
          type: string
        projectId:
          type: integer
        projectName:
          type: string
        content:
          type: string
        isRead:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - senderId
        - content
        - isRead
        - createdAt
    SendMessageRequest:
      type: object
      properties:
        recipientId:
          type: integer
        projectId:
          type: integer
        content:
          type: string
      required:
        - content
...