# lib/api-zod/src/generated/types/refreshTokenRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * HAIROU BTP Management Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface RefreshTokenRequest {
  refreshToken: string;
}